first succesful build after upgrades and rename
