+++ src/OpenTelemetry.Exporter.GoogleCloud/Implementation/ActivityExtensions.cs
@@ -1,7 +1,11 @@
 // Copyright The OpenTelemetry Authors
 // SPDX-License-Identifier: Apache-2.0
 
+using System.Diagnostics;
 using System.Globalization;
+using Google.Cloud.Trace.V2;
+using Google.Protobuf.WellKnownTypes;
+using OpenTelemetry.Trace;
 
 namespace OpenTelemetry.Exporter.GoogleCloud.Implementation;
