+++ src/OpenTelemetry.Exporter.GoogleCloud/Implementation/GoogleCloudResourceUtils.cs
@@ -1,6 +1,8 @@
 // Copyright The OpenTelemetry Authors
 // SPDX-License-Identifier: Apache-2.0
 
+using Google.Api;
+
 namespace OpenTelemetry.Exporter.GoogleCloud.Implementation;
 
 /// <summary>
