+++ src/OpenTelemetry.Exporter.GoogleCloud/Implementation/ResourceExtensions.cs
@@ -1,6 +1,9 @@
 // Copyright The OpenTelemetry Authors
 // SPDX-License-Identifier: Apache-2.0
 
+using Google.Cloud.Trace.V2;
+using OpenTelemetry.Resources;
+
 namespace OpenTelemetry.Exporter.GoogleCloud.Implementation;
 
 internal static class ResourceExtensions
