+++ src/OpenTelemetry.Exporter.GoogleCloud/README.md
@@ -1,4 +1,4 @@
-# Stackdriver Exporter for OpenTelemetry .NET
+# Google Cloud Exporter for OpenTelemetry .NET
 
 | Status        |           |
 | ------------- |-----------|
@@ -18,11 +18,11 @@ GOOGLE_APPLICATION_CREDENTIALS. When you run on [GAE][GAE], [GKE][GKE] or
 locally with gcloud sdk installed - this is typically the case. There is also a
 constructor for specifying path to the service account credential.
 
-1. Add [Stackdriver Exporter
-   package][OpenTelemetry-exporter-stackdriver-myget-url] reference.
-2. Enable [Stackdriver Trace][stackdriver-trace-setup] API.
-3. Enable [Stackdriver Monitoring][stackdriver-monitoring-setup] API.
-4. Instantiate a new instance of `StackdriverExporter` with your Google Cloud's
+1. Add [Google Cloud Exporter
+   package][OpenTelemetry-exporter-googlecloud-myget-url] reference.
+2. Enable [Google Cloud Trace][googlecloud-trace-setup] API.
+3. Enable [Google Cloud Monitoring][googlecloud-monitoring-setup] API.
+4. Instantiate a new instance of `GoogleCloudExporter` with your Google Cloud's
    ProjectId
 
 ## Installation
@@ -55,18 +55,18 @@ using (tracer.StartActiveSpan("/getuser", out TelemetrySpan span))
 
 ## References
 
-* [stackdriver-trace-setup](https://cloud.google.com/trace/docs/setup/)
-* [stackdriver-monitoring-setup](https://cloud.google.com/monitoring/api/enable-api)
+* [googlecloud-trace-setup](https://cloud.google.com/trace/docs/setup/)
+* [googlecloud-monitoring-setup](https://cloud.google.com/monitoring/api/enable-api)
 * [GAE](https://cloud.google.com/appengine/docs/flexible/dotnet/quickstart)
 * [GKE](https://codelabs.developers.google.com/codelabs/cloud-kubernetes-aspnetcore/index.html)
 * [gcp-auth](https://cloud.google.com/docs/authentication/getting-started)
 
-[stackdriver-trace-setup]: https://cloud.google.com/trace/docs/setup/
-[stackdriver-monitoring-setup]:
+[googlecloud-trace-setup]: https://cloud.google.com/trace/docs/setup/
+[googlecloud-monitoring-setup]:
     https://cloud.google.com/monitoring/api/enable-api
 [GAE]: https://cloud.google.com/appengine/docs/flexible/dotnet/quickstart
 [GKE]:
     https://codelabs.developers.google.com/codelabs/cloud-kubernetes-aspnetcore/index.html
 [gcp-auth]: https://cloud.google.com/docs/authentication/getting-started
-[OpenTelemetry-exporter-stackdriver-myget-url]:
-    https://www.nuget.org/packages/OpenTelemetry.Exporter.Stackdriver
+[OpenTelemetry-exporter-googlecloud-myget-url]:
+    https://www.nuget.org/packages/OpenTelemetry.Exporter.GoogleCloud
