+++ src/OpenTelemetry.Exporter.GoogleCloud/TracerProviderBuilderExtensions.cs
@@ -2,16 +2,17 @@
 // SPDX-License-Identifier: Apache-2.0
 
 using OpenTelemetry.Internal;
+using OpenTelemetry.Trace;
 
 namespace OpenTelemetry.Exporter.GoogleCloud;
 
 /// <summary>
-/// Extension methods to simplify registering a Stackdriver exporter.
+/// Extension methods to simplify registering a Google Cloud exporter.
 /// </summary>
 public static class TracerProviderBuilderExtensions
 {
     /// <summary>
-    /// Registers a Stackdriver exporter that will receive <see cref="System.Diagnostics.Activity"/> instances.
+    /// Registers a Google Cloud exporter that will receive <see cref="System.Diagnostics.Activity"/> instances.
     /// </summary>
     /// <param name="builder"><see cref="TracerProviderBuilder"/> builder to use.</param>
     /// <param name="projectId">Project ID to send telemetry to.</param>
