+++ test/OpenTelemetry.Exporter.GoogleCloud.Tests/StackdriverExporterTests.cs
@@ -2,12 +2,12 @@
 // SPDX-License-Identifier: Apache-2.0
 
 using System.Diagnostics;
-using OpenTelemetry.Exporter.Stackdriver.Implementation;
+using OpenTelemetry.Exporter.GoogleCloud.Implementation;
 using OpenTelemetry.Resources;
 using OpenTelemetry.Trace;
 using Xunit;
 
-namespace OpenTelemetry.Exporter.Stackdriver.Tests;
+namespace OpenTelemetry.Exporter.GoogleCloud.Tests;
 
 public class StackdriverExporterTests
 {
@@ -50,7 +50,7 @@ public class StackdriverExporterTests
         var openTelemetrySdk = Sdk.CreateTracerProviderBuilder()
             .AddSource(ActivitySourceName)
             .AddProcessor(testActivityProcessor)
-            .UseStackdriverExporter("test").Build();
+            .UseGoogleCloudExporter("test").Build();
 
         var source = new ActivitySource(ActivitySourceName);
         var activity = source.StartActivity("Test Activity");
@@ -66,7 +66,7 @@ public class StackdriverExporterTests
         const string ActivitySourceName = "stackdriver.test";
 
         var traceClient = new TestTraceServiceClient(throwException: false);
-        var activityExporter = new StackdriverTraceExporter("test_project", traceClient);
+        var activityExporter = new GoogleCloudTraceExporter("test_project", traceClient);
 
         var openTelemetrySdk = Sdk.CreateTracerProviderBuilder()
             .AddSource(ActivitySourceName)
@@ -91,7 +91,7 @@ public class StackdriverExporterTests
         const string ActivitySourceName = "stackdriver.test";
         var source = new ActivitySource(ActivitySourceName);
         var traceClient = new TestTraceServiceClient(throwException: true);
-        var activityExporter = new StackdriverTraceExporter("test", traceClient);
+        var activityExporter = new GoogleCloudTraceExporter("test", traceClient);
 
         var processor = new BatchActivityExportProcessor(new InMemoryExporter<Activity>(exportedItems));
 
@@ -127,7 +127,7 @@ public class StackdriverExporterTests
         const string ActivitySourceName = "stackdriver.test";
         var source = new ActivitySource(ActivitySourceName);
         var traceClient = new TestTraceServiceClient(throwException: false);
-        var activityExporter = new StackdriverTraceExporter("test", traceClient);
+        var activityExporter = new GoogleCloudTraceExporter("test", traceClient);
 
         var processor = new BatchActivityExportProcessor(new InMemoryExporter<Activity>(exportedItems));
